Cleanup and fixing a few oopsies

+++ entrypoint.sh
@@ -5,8 +5,17 @@ adduser -S -h /srv/sabredav -D -H -s /sbin/nologin -u $PUID -G unit unit
 
 chown -R $PUID:$PGID /srv/sabredav /data
 
-/opt/sbin/unitd --user unit --group unit
-curl --data-binary @/config/unit/config.json -X PUT --unix-socket /socket/control/control.unit.sock http://localhost/config/
-pkill unitd
+
+if [ -e /socket/sabredav/sabredav.sock ]
+then
+	rm /socket/sabredav/sabredav.sock
+fi
+if [ "$AUTOCONFIG" == "YES" ]
+then
+	/opt/sbin/unitd
+	curl --data-binary @/config/unit/config.json -X PUT --unix-socket /socket/control/control.unit.sock http://localhost/config/
+	pkill unitd
+fi
+
 rm /socket/sabredav/sabredav.sock
-exec /opt/sbin/unitd --no-daemon --user unit --group unit
+exec /opt/sbin/unitd --no-daemon
